Map Explorer and Map Visuals have split!
